# Conflicts:
#	document/src/test/java/com/ritense/document/service/impl/JsonSchemaDocumentServiceTest.java
#	zgw/build.gradle
#	zgw/catalogi-api/src/main/kotlin/com/ritense/catalogiapi/CatalogiApiAutoConfiguration.kt
#	zgw/catalogi-api/src/main/kotlin/com/ritense/catalogiapi/CatalogiApiPlugin.kt
#	zgw/catalogi-api/src/main/kotlin/com/ritense/catalogiapi/client/CatalogiApiClient.kt
#	zgw/catalogi-api/src/main/kotlin/com/ritense/catalogiapi/service/CatalogiService.kt
#	zgw/documenten-api/src/main/kotlin/com/ritense/documentenapi/security/DocumentenApiHttpSecurityConfigurer.kt
#	zgw/documenten-api/src/main/kotlin/com/ritense/documentenapi/web/rest/DocumentenApiResource.kt
#	zgw/documenten-api/src/test/kotlin/com/ritense/documentenapi/DocumentenApiPluginFactoryTest.kt
#	zgw/documenten-api/src/test/kotlin/com/ritense/documentenapi/DocumentenApiPluginTest.kt
#	zgw/zaken-api/src/main/kotlin/com/ritense/zakenapi/config/ZakenApiAutoConfiguration.kt

+++ plugin/src/main/kotlin/com/ritense/plugin/service/PluginService.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2015-2023 Ritense BV, the Netherlands.
+ * Copyright 2015-2024 Ritense BV, the Netherlands.
  *
  * Licensed under EUPL, Version 1.2 (the "License");
  * you may not use this file except in compliance with the License.
@@ -27,7 +27,6 @@ import com.ritense.plugin.annotation.PluginActionProperty
 import com.ritense.plugin.annotation.PluginCategory
 import com.ritense.plugin.annotation.PluginEvent
 import com.ritense.plugin.autodeployment.PluginAutoDeploymentDto
-import com.ritense.plugin.domain.ActivityType
 import com.ritense.plugin.domain.EventType
 import com.ritense.plugin.domain.PluginActionDefinition
 import com.ritense.plugin.domain.PluginConfiguration
@@ -35,6 +34,9 @@ import com.ritense.plugin.domain.PluginConfigurationId
 import com.ritense.plugin.domain.PluginDefinition
 import com.ritense.plugin.domain.PluginProcessLink
 import com.ritense.plugin.domain.PluginProcessLinkId
+import com.ritense.plugin.domain.PluginProperty
+import com.ritense.plugin.events.PluginConfigurationDeletedEvent
+import com.ritense.plugin.events.PluginConfigurationIdUpdatedEvent
 import com.ritense.plugin.exception.PluginEventInvocationException
 import com.ritense.plugin.exception.PluginPropertyParseException
 import com.ritense.plugin.exception.PluginPropertyRequiredException
@@ -47,18 +49,20 @@ import com.ritense.plugin.web.rest.request.PluginProcessLinkCreateDto
 import com.ritense.plugin.web.rest.request.PluginProcessLinkUpdateDto
 import com.ritense.plugin.web.rest.result.PluginActionDefinitionDto
 import com.ritense.plugin.web.rest.result.PluginProcessLinkResultDto
+import com.ritense.processlink.domain.ActivityTypeWithEventName
 import com.ritense.valueresolver.ValueResolverService
+import jakarta.validation.ConstraintViolationException
+import jakarta.validation.ValidationException
+import jakarta.validation.Validator
 import mu.KotlinLogging
 import org.camunda.bpm.engine.delegate.DelegateExecution
 import org.camunda.bpm.engine.delegate.DelegateTask
+import org.springframework.context.ApplicationEventPublisher
 import org.springframework.data.repository.findByIdOrNull
 import org.springframework.transaction.annotation.Transactional
 import java.lang.reflect.Method
 import java.lang.reflect.Parameter
 import java.util.UUID
-import javax.validation.ConstraintViolationException
-import javax.validation.ValidationException
-import javax.validation.Validator
 import kotlin.reflect.full.findAnnotation
 import kotlin.reflect.full.functions
 
@@ -73,6 +77,8 @@ class PluginService(
     private val valueResolverService: ValueResolverService,
     private val pluginConfigurationSearchRepository: PluginConfigurationSearchRepository,
     private val validator: Validator,
+    private val applicationEventPublisher: ApplicationEventPublisher,
+    private val encryptionService: EncryptionService,
 ) {
 
     fun getObjectMapper(): ObjectMapper {
@@ -115,12 +121,23 @@ class PluginService(
         properties: ObjectNode,
         pluginDefinitionKey: String
     ): PluginConfiguration {
-        val pluginDefinition = pluginDefinitionRepository.getById(pluginDefinitionKey)
+        val pluginDefinition = pluginDefinitionRepository.getReferenceById(pluginDefinitionKey)
         validateProperties(properties, pluginDefinition)
 
-        val pluginConfiguration = pluginConfigurationRepository.save(
-            PluginConfiguration(id, title, properties, pluginDefinition)
+        if (pluginConfigurationRepository.existsById(id)) {
+            throw IllegalStateException("Failed to create plugin. Plugin ID '${id.id}' is already used by another plugin.")
+        }
+
+        val pluginConfigurationToBeSaved = PluginConfiguration(
+            id,
+            title,
+            properties,
+            pluginDefinition,
+            encryptionService,
+            objectMapper
         )
+
+        val pluginConfiguration = pluginConfigurationRepository.save(pluginConfigurationToBeSaved)
 
         try {
             pluginConfiguration.runAllPluginEvents(EventType.CREATE)
@@ -134,30 +151,41 @@ class PluginService(
     }
 
     fun deployPluginConfigurations(deploymentDto: PluginAutoDeploymentDto) {
-        val plugin: PluginConfiguration
-        val pluginDefinition = pluginDefinitionRepository.getById(deploymentDto.pluginDefinitionKey)
-        if (deploymentDto.id != null && pluginConfigurationRepository.existsById(
-                PluginConfigurationId.existingId(deploymentDto.id)
-            )
-        ) {
-            deletePluginConfiguration(PluginConfigurationId.existingId(deploymentDto.id))
-        }
+        val pluginConfiguration: PluginConfiguration
+        var oldConfiguration: PluginConfiguration? = null
+        var action: EventType = EventType.CREATE
+
+        val pluginDefinition = pluginDefinitionRepository.getReferenceById(deploymentDto.pluginDefinitionKey)
         val resolvedProperties = resolveProperties(deploymentDto.properties)
         validateProperties(resolvedProperties, pluginDefinition)
-        val pluginId = pluginConfigurationRepository.saveAndFlush(
-            PluginConfiguration(
-                deploymentDto.id?.let { PluginConfigurationId.existingId(it) } ?: PluginConfigurationId.newId(),
-                deploymentDto.title,
-                resolvedProperties,
-                pluginDefinition
-            )
-        ).id
-        plugin = pluginConfigurationRepository.findById(pluginId).orElseThrow()
+
+        deploymentDto.id?.let {
+            oldConfiguration = pluginConfigurationRepository.findByIdOrNull(PluginConfigurationId.existingId(deploymentDto.id))
+            if (oldConfiguration != null) {
+                action = EventType.UPDATE
+            }
+        }
+
+        pluginConfiguration = PluginConfiguration(
+            deploymentDto.id?.let { PluginConfigurationId.existingId(it) } ?: PluginConfigurationId.newId(),
+            deploymentDto.title,
+            resolvedProperties,
+            pluginDefinition,
+            encryptionService,
+            objectMapper
+        )
+
+        pluginConfigurationRepository.saveAndFlush(pluginConfiguration)
         try {
-            plugin.runAllPluginEvents(EventType.CREATE)
+            pluginConfiguration.runAllPluginEvents(action)
         } catch (e: Exception) {
-            pluginConfigurationRepository.deleteById(plugin.id)
-            throw PluginEventInvocationException(plugin, e)
+            if (oldConfiguration != null) {
+                //restore old configuration
+                pluginConfigurationRepository.save(oldConfiguration)
+            } else {
+                pluginConfigurationRepository.deleteById(pluginConfiguration.id)
+            }
+            throw PluginEventInvocationException(pluginConfiguration, e)
         }
     }
 
@@ -231,15 +259,18 @@ class PluginService(
                 } catch (e: Exception) {
                     logger.warn { "Failed to run events on plugin ${it.title} with id ${it.id.id}" }
                 }
+
+                pluginConfigurationRepository.deleteById(pluginConfigurationId)
+
+                val event = PluginConfigurationDeletedEvent(it)
+                applicationEventPublisher.publishEvent(event)
             }
             ?: logger.warn { "Plugin configuration with Id: [$pluginConfigurationId] was not found." }
-
-        pluginConfigurationRepository.deleteById(pluginConfigurationId)
     }
 
     fun getPluginDefinitionActions(
         pluginDefinitionKey: String,
-        activityType: ActivityType?
+        activityType: ActivityTypeWithEventName?
     ): List<PluginActionDefinitionDto> {
         val actions = if (activityType == null)
             pluginActionDefinitionRepository.findByIdPluginDefinitionKey(pluginDefinitionKey)
@@ -262,7 +293,7 @@ class PluginService(
     fun processLinkExists(
         pluginConfigurationId: PluginConfigurationId,
         activityId: String,
-        activityType: ActivityType
+        activityType: ActivityTypeWithEventName
     ): Boolean {
         return pluginProcessLinkRepository
             .findByPluginConfigurationIdAndActivityIdAndActivityType(
@@ -304,7 +335,7 @@ class PluginService(
             actionProperties = processLink.actionProperties,
             pluginConfigurationId = PluginConfigurationId.existingId(processLink.pluginConfigurationId),
             pluginActionDefinitionKey = processLink.pluginActionDefinitionKey,
-            activityType = ActivityType.fromValue(processLink.activityType.value).mapOldActivityTypeToCurrent()
+            activityType = processLink.activityType
         )
         pluginProcessLinkRepository.save(newProcessLink)
     }
@@ -356,27 +387,29 @@ class PluginService(
         if (newPluginConfigurationId == oldPluginConfigurationId) {
             return oldPluginConfiguration
         }
+        if (pluginConfigurationRepository.existsById(newPluginConfigurationId)) {
+            throw IllegalStateException("Failed to update plugin. Plugin ID '${newPluginConfigurationId.id}' is already used by another plugin.")
+        }
         pluginConfigurationRepository.deleteById(oldPluginConfigurationId)
         val newPluginConfiguration = pluginConfigurationRepository.save(
             PluginConfiguration(
                 newPluginConfigurationId,
                 oldPluginConfiguration.title,
-                oldPluginConfiguration.rawProperties,
-                oldPluginConfiguration.pluginDefinition
+                oldPluginConfiguration.properties,
+                oldPluginConfiguration.pluginDefinition,
+                encryptionService,
+                objectMapper
             )
         )
-        val processLinks = pluginProcessLinkRepository.findByPluginConfigurationId(oldPluginConfigurationId)
-            .map { it.copy(pluginConfigurationId = newPluginConfigurationId) }
-        pluginProcessLinkRepository.saveAll(processLinks)
-        val configurations = pluginConfigurationRepository.findAll()
-        configurations.forEach { configuration ->
-            configuration.rawProperties?.fields()?.forEachRemaining { property ->
-                if (property.value.textValue() == oldPluginConfigurationId.id.toString()) {
-                    property.setValue(TextNode.valueOf(newPluginConfigurationId.id.toString()))
-                }
-            }
-        }
-        pluginConfigurationRepository.saveAll(configurations)
+
+        val event = PluginConfigurationIdUpdatedEvent(
+            newPluginConfigurationId.id,
+            oldPluginConfigurationId.id,
+            newPluginConfiguration
+        )
+
+        applicationEventPublisher.publishEvent(event)
+
         return newPluginConfiguration
     }
 
@@ -436,7 +469,7 @@ class PluginService(
         }.toMap()
 
         // We want to process all placeholder values together to improve performance if external sources are needed.
-        val placeHolderValueMap = method.parameters.filter { param ->
+        val resolvedValueMap = method.parameters.filter { param ->
             param.isAnnotationPresent(PluginActionProperty::class.java)
         }.mapNotNull { param ->
             param to actionProperties.get(param.name)
@@ -449,7 +482,7 @@ class PluginService(
                 valueResolverService.resolveValues(execution.processInstanceId, execution, values.toList())
             }
 
-        return mapActionParamValues(paramValues, placeHolderValueMap)
+        return mapActionParamValues(paramValues, resolvedValueMap)
     }
 
     private fun resolveActionParamValues(
@@ -470,7 +503,7 @@ class PluginService(
         }.toMap()
 
         // We want to process all placeholder values together to improve performance if external sources are needed.
-        val placeHolderValueMap = method.parameters.filter { param ->
+        val resolvedValueMap = method.parameters.filter { param ->
             param.isAnnotationPresent(PluginActionProperty::class.java)
         }.mapNotNull { param ->
             param to actionProperties.get(param.name)
@@ -483,19 +516,18 @@ class PluginService(
                 valueResolverService.resolveValues(task.execution.processInstanceId, task.execution, values.toList())
             }
 
-        return mapActionParamValues(paramValues, placeHolderValueMap)
+        return mapActionParamValues(paramValues, resolvedValueMap)
     }
 
     private fun mapActionParamValues(
         paramValues: Map<Parameter, JsonNode>,
-        placeHolderValueMap: Map<String, Any>
+        resolvedValueMap: Map<String, Any?>
     ): Map<Parameter, Any> {
         return paramValues.mapValues { (param, value) ->
             if (value.isTextual) {
-                val placeHolderValueAsString = placeHolderValueMap[value.textValue()]
-                if (placeHolderValueAsString != null) {
+                if (resolvedValueMap.containsKey(value.textValue())) {
                     objectMapper.convertValue(
-                        placeHolderValueAsString,
+                        resolvedValueMap[value.textValue()],
                         objectMapper.constructType(param.parameterizedType)
                     )
                 } else {
@@ -520,7 +552,7 @@ class PluginService(
     }
 
     fun createInstance(pluginConfigurationId: PluginConfigurationId): Any {
-        val configuration = pluginConfigurationRepository.getById(pluginConfigurationId)
+        val configuration = pluginConfigurationRepository.getReferenceById(pluginConfigurationId)
         return createInstance(configuration)
     }
 
@@ -534,13 +566,6 @@ class PluginService(
         val pluginConfiguration = findPluginConfiguration(clazz, configurationFilter)
 
         return pluginConfiguration?.let { createInstance(it) as T }
-    }
-
-    fun <T> findPluginConfiguration(clazz: Class<T>, configurationFilter: (JsonNode) -> Boolean): PluginConfiguration? {
-        val annotation = clazz.getAnnotation(Plugin::class.java)
-            ?: throw IllegalArgumentException("Requested plugin for class ${clazz.name}, but class is not annotated as plugin")
-
-        return findPluginConfiguration(annotation.key, configurationFilter)
     }
 
     private fun getActionMethod(
@@ -568,22 +593,7 @@ class PluginService(
                 errors.add(PluginPropertyRequiredException(pluginProperty.fieldName, pluginDefinition.title))
             } else {
                 try {
-                    val propertyClass = Class.forName(pluginProperty.fieldType)
-                    val propertyClassIsPlugin = propertyClass.isAnnotationPresent(Plugin::class.java)
-                    val propertyClassIsPluginCategory = propertyClass.isAnnotationPresent(PluginCategory::class.java)
-                    if (propertyClassIsPlugin || propertyClassIsPluginCategory) {
-                        val propertyConfigurationId =
-                            PluginConfigurationId.existingId(UUID.fromString(propertyNode.textValue()))
-                        val propertyConfiguration = pluginConfigurationRepository.findById(propertyConfigurationId)
-                        assert(propertyConfiguration.isPresent) { "Plugin configuration with id ${propertyConfigurationId.id} does not exist!" }
-                    } else {
-                        val propertyValue = objectMapper.treeToValue(propertyNode, propertyClass)
-                        val validationErrors =
-                            validator.validateValue(pluginClass, pluginProperty.fieldName, propertyValue)
-                        if (validationErrors.isNotEmpty()) {
-                            throw ConstraintViolationException(validationErrors)
-                        }
-                    }
+                    validateProperty(pluginProperty, propertyNode, pluginClass)
                 } catch (e: Exception) {
                     errors.add(PluginPropertyParseException(pluginProperty.fieldName, pluginDefinition.title, e))
                 }
@@ -596,12 +606,52 @@ class PluginService(
         }
     }
 
+    @Throws(ConstraintViolationException::class)
+    private fun validateProperty(pluginProperty: PluginProperty, propertyNode: JsonNode?, pluginClass: Class<*>) {
+        val propertyClass = Class.forName(pluginProperty.fieldType)
+        val propertyClassIsPlugin = propertyClass.isAnnotationPresent(Plugin::class.java)
+        val propertyClassIsPluginCategory = propertyClass.isAnnotationPresent(PluginCategory::class.java)
+        if (propertyClassIsPlugin || propertyClassIsPluginCategory) {
+            val propertyConfigurationId =
+                PluginConfigurationId.existingId(UUID.fromString(propertyNode?.textValue()))
+            val propertyConfiguration = pluginConfigurationRepository.findById(propertyConfigurationId)
+            assert(propertyConfiguration.isPresent) { "Plugin configuration with id ${propertyConfigurationId.id} does not exist!" }
+        } else {
+            val propertyValue = objectMapper.treeToValue(propertyNode, propertyClass)
+            val validationErrors =
+                validator.validateValue(pluginClass, pluginProperty.fieldName, propertyValue)
+            if (validationErrors.isNotEmpty()) {
+                throw ConstraintViolationException(validationErrors)
+            }
+        }
+    }
+
+    fun <T> findPluginConfiguration(clazz: Class<T>, configurationFilter: (JsonNode) -> Boolean): PluginConfiguration? {
+        return findPluginConfigurations(clazz, configurationFilter)
+            .firstOrNull()
+    }
+
     fun findPluginConfiguration(
         pluginDefinitionKey: String,
         filter: (JsonNode) -> Boolean
     ): PluginConfiguration? {
+        return findPluginConfigurations(pluginDefinitionKey, filter)
+            .firstOrNull()
+    }
+
+    fun <T> findPluginConfigurations(clazz: Class<T>, filter: (JsonNode) -> Boolean = { true }): List<PluginConfiguration> {
+        val annotation = clazz.getAnnotation(Plugin::class.java)
+            ?: throw IllegalArgumentException("Requested plugin for class ${clazz.name}, but class is not annotated as plugin")
+
+        return findPluginConfigurations(annotation.key, filter)
+    }
+
+    private fun findPluginConfigurations(
+        pluginDefinitionKey: String,
+        filter: (JsonNode) -> Boolean
+    ): List<PluginConfiguration> {
         val configurations = pluginConfigurationRepository.findByPluginDefinitionKey(pluginDefinitionKey)
-        return configurations.firstOrNull { config ->
+        return configurations.filter { config ->
             val configProperties = config.properties
             configProperties != null && filter(configProperties)
         }
@@ -628,7 +678,7 @@ class PluginService(
     }
 
     fun getPluginDefinitionActionsByActivityType(activityType: String): List<PluginActionDefinition> {
-        return pluginActionDefinitionRepository.findByActivityTypes(ActivityType.fromValue(activityType))
+        return pluginActionDefinitionRepository.findByActivityTypes(ActivityTypeWithEventName.fromValue(activityType))
     }
 
     companion object {
